move fns to translate_align
