+++ src/sr2silo/process/__init__.py
@@ -15,7 +15,10 @@ from sr2silo.process.interface import (
     GeneSet,
 )
 from sr2silo.process.merge import pair_normalize_reads
-from sr2silo.process.translate_align import translate_nextclade, nuc_to_aa_alignment
+from sr2silo.process.translate_align import (
+                  translate_nextclade,
+                  nuc_to_aa_alignment
+)
 from sr2silo.process.convert import get_gene_set_from_ref
 
 __all__ = [
